+++ lib/irie/extensions/conversion/nil_params.rb
@@ -12,13 +12,15 @@ module Irie
 
         # Converts request param value(s) 'NULL', 'null', and 'nil' to nil.
         def convert_param(param_name, param_value_or_values)
-          logger.debug("Irie::Extensions::Conversion::NilParams.convert_param(#{param_name.inspect}, #{param_value_or_values.inspect})") if ::Irie.debug?
+          ::Irie.logger.debug("[Irie] Irie::Extensions::Conversion::NilParams.convert_param(#{param_name.inspect}, #{param_value_or_values.inspect})") if ::Irie.debug?
           param_value_or_values = super if defined?(super)
           if param_value_or_values.is_a? Array
-            param_value_or_values.map{|v| v && NILS.include?(v) ? nil : v }
+            result = param_value_or_values.map{|v| v && NILS.include?(v) ? nil : v }
           else
-            param_value_or_values && NILS.include?(param_value_or_values) ? nil : param_value_or_values
+            result = param_value_or_values && NILS.include?(param_value_or_values) ? nil : param_value_or_values
           end
+          ::Irie.logger.debug("[Irie] Irie::Extensions::Conversion::NilParams.convert_param: result: #{result.inspect}") if ::Irie.verbose?
+          result
         end
       
       end