+++ lib/irie/extensions/count.rb
@@ -10,10 +10,15 @@ module Irie
       end
       
       def index(options={}, &block)
-        logger.debug("Irie::Extensions::Count.index") if ::Irie.debug?
+        ::Irie.logger.debug("[Irie] Irie::Extensions::Count.index") if ::Irie.debug?
         return super(options, &block) unless aliased_param_present?(:count)
         @count = collection.count
-        return respond_to?(:autorender_count, true) ? autorender_count(options, &block) : super(options, &block)
+
+        ::Irie.logger.debug("[Irie] Irie::Extensions::Count.index: count: #{@count}") if ::Irie.debug?
+
+        result = respond_to?(:autorender_count, true) ? autorender_count(options, &block) : super(options, &block)
+        ::Irie.logger.debug("[Irie] Irie::Extensions::Count.index: result: #{result.inspect}") if ::Irie.verbose?
+        result
       end
 
     end