+++ lib/irie/extensions/limit.rb
@@ -12,11 +12,15 @@ module Irie
       protected
 
       def collection
-        logger.debug("Irie::Extensions::Limit.collection") if ::Irie.debug?
+        ::Irie.logger.debug("[Irie] Irie::Extensions::Limit.collection") if ::Irie.debug?
         object = super
-        aliased_params(:limit).each {|param_value| object = object.limit(param_value)}
+        ::Irie.logger.debug("[Irie] Irie::Extensions::Limit.collection: starting after super with object=#{object.inspect}") if ::Irie.verbose?
+        limit_params = aliased_params(:limit)
+        ::Irie.logger.debug("[Irie] Irie::Extensions::Limit.collection: limit_params=#{limit_params.inspect}") if ::Irie.debug?
+        limit_params.each {|param_value| object = object.limit(param_value.to_i)}
+        ::Irie.logger.debug("[Irie] Irie::Extensions::Limit.collection: after limits with object=#{object.inspect}") if ::Irie.verbose?
 
-        logger.debug("Irie::Extensions::Limit.collection: relation.to_sql so far: #{object.to_sql}") if ::Irie.debug? && object.respond_to?(:to_sql)
+        ::Irie.logger.debug("[Irie] Irie::Extensions::Limit.collection: relation.to_sql so far: #{object.to_sql}") if ::Irie.debug? && object.respond_to?(:to_sql)
 
         set_collection_ivar object
       end