+++ lib/irie/extensions/paging/autorender_page_count.rb
@@ -9,8 +9,10 @@ module Irie
         protected
 
         def autorender_page_count(options={}, &block)
-          logger.debug("Irie::Extensions::Paging::AutorenderPageCount.autorender_page_count") if ::Irie.debug?
-          render request.format.symbol => { page_count: @page_count }, status: 200, layout: false
+          ::Irie.logger.debug("[Irie] Irie::Extensions::Paging::AutorenderPageCount.autorender_page_count: page_count: #{@page_count.inspect}") if ::Irie.debug?
+          result = render(request.format.symbol => { page_count: @page_count }, status: 200, layout: false)
+          ::Irie.logger.debug("[Irie] Irie::Extensions::Paging::AutorenderPageCount.autorender_page_count: result: #{result.inspect}") if ::Irie.verbose?
+          result
         end
         
       end