+++ lib/irie/extensions/query_includes.rb
@@ -62,12 +62,16 @@ module Irie
         
         this_includes = self.action_to_query_includes[params[:action].to_sym] || self.all_action_query_includes
         if this_includes && this_includes.size > 0
+          ::Irie.logger.debug("[Irie] Irie::Extensions::QueryIncludes.collection: calling .includes(*#{this_includes.inspect})") if ::Irie.debug?
           object = object.includes(*this_includes)
         else
+          ::Irie.logger.debug("[Irie] Irie::Extensions::QueryIncludes.collection: not calling .includes") if ::Irie.debug?
           object
         end
 
-        logger.debug("Irie::Extensions::QueryIncludes.collection: relation.to_sql so far: #{object.to_sql}") if ::Irie.debug? && object.respond_to?(:to_sql)
+        ::Irie.logger.debug("[Irie] Irie::Extensions::QueryIncludes.collection: relation.to_sql so far: #{object.to_sql}") if ::Irie.debug? && object.respond_to?(:to_sql)
+
+        ::Irie.logger.debug("[Irie] Irie::Extensions::QueryIncludes.collection: object = #{object.inspect}") if ::Irie.verbose?
 
         set_collection_ivar object
       end
@@ -75,19 +79,22 @@ module Irie
       def resource
         cached = get_resource_ivar
         if cached
-          logger.debug("Irie::Extensions::QueryIncludes.resource returning cached resource") if ::Irie.debug?
+          ::Irie.logger.debug("[Irie] Irie::Extensions::QueryIncludes.resource: returning cached resource") if ::Irie.debug?
           return cached
         end
         logger.debug("Irie::Extensions::QueryIncludes.resource") if ::Irie.debug?
+
         this_includes = self.action_to_query_includes[params[:action].to_sym] || self.all_action_query_includes
         if this_includes && this_includes.size > 0
           # can return the model class, so won't call bang (includes!) method
           object = end_of_association_chain.includes(*this_includes)
 
-          logger.debug("Irie::Extensions::QueryIncludes.resource: end_of_association_chain.to_sql: #{object.to_sql}") if ::Irie.debug? && object.respond_to?(:to_sql)
+          ::Irie.logger.debug("[Irie] Irie::Extensions::QueryIncludes.resource: end_of_association_chain.to_sql: #{object.to_sql}") if ::Irie.debug? && object.respond_to?(:to_sql)
 
           set_resource_ivar object.send(method_for_find, params[:id])
         else
+          ::Irie.logger.debug("[Irie] Irie::Extensions::QueryIncludes.resource") if ::Irie.debug?
+
           super
         end
       end