+++ test/dummy/config/initializers/0_logging_to_stdout.rb
@@ -1,0 +1,3 @@
+Rails.logger = Logger.new(STDOUT)
+Rails.logger.level = Logger::WARN
+#ActiveRecord::Base.logger = Logger.new(STDOUT)